Fix two reset_network swedish translations
The reset_network_title translation meant "Network settings HAS BEEN
restored" which is confusing as it is a command to restore network
settings.

The reset_network_complete_toast translation meant "Network settings
has NOT been restored" which is wrong since it is used to indicate
that the settings actually were restored.

Change-Id: I0128b873172425d5e101f8541c76a24ebd94b24c