Allow unlocked USB data access (2/2)
by @KreAch3R

This bypasses the marshmallow "regression" of blocking and resetting USB data access
after disconnecting. It also allows saving the default configuration and it survives a reboot.
Originally, this "feature" was added by Google for security reasons
so it should be advised to use the below setting under caution.

Also:
 * Adds an "Advanced" section under Storage and USB
 * Moves Usb Configuration from Development settings
 * Secured lockscreen disables the option

Change-Id: Ia951d6d49ff91d4e28553205fa5a844620f2cb70